refactor: move generators into generators.py
